+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All Notable changes to `Laravel Git Hooks` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## NEXT - YYYY-MM-DD
+## NEXT
 
 ### Added
-- Nothing
+- prepare-commit-msg git hook
+- enable package discovery for Laravel 5.5
+- added post-checkout git hook #9
+- add install dependencies command #10
 
 ### Deprecated
 - Nothing
@@ -20,3 +23,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Security
 - Nothing
+
+## [v0.3.1]
+
+Begin tracking changelog :)